+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.88] - 2024-06-12
+### Added
+* `ColorExtensions.ToNegative` to flip a color
+* commented-out methods within `RectTransformExtensions`
+* `WorldToScreenPoint` to `ScreenUtils`
+### Fixed
+* `EditorUtils` not saving after drawing properties
+* `ScenesInBuildManager` being in the wrong namespace
+* `SchedulingEventManager`'s prefab naming violation
+
 ## [0.0.87] - 2024-05-28
 ### Added
 * `MathKoko.RemapFrom01`